refactor: model 폴더 생성 후 해당 파일 이동
